Extracted data modification logic from output serializers to separate mapper
