+++ core/server/api/v2/utils/serializers/output/pages.js
@@ -1,22 +1,5 @@
 const debug = require('ghost-ignition').debug('api:v2:utils:serializers:output:pages');
-const url = require('./utils/url');
-const date = require('./utils/date');
-const utils = require('../../');
-
-const mapPage = (model, frame) => {
-    const jsonModel = model.toJSON(frame.options);
-    url.forPost(model.id, jsonModel, frame.options);
-
-    if (utils.isContentAPI(frame)) {
-        ['created_at', 'updated_at', 'published_at'].forEach((field) => {
-            if (jsonModel[field]) {
-                jsonModel[field] = date.format(jsonModel[field]);
-            }
-        });
-    }
-
-    return jsonModel;
-};
+const mapper = require('./utils/mapper');
 
 module.exports = {
     all(models, apiConfig, frame) {
@@ -24,7 +7,7 @@ module.exports = {
 
         if (models.meta) {
             frame.response = {
-                pages: models.data.map(model => mapPage(model, frame)),
+                pages: models.data.map(model => mapper.mapPage(model, frame)),
                 meta: models.meta
             };
 
@@ -32,7 +15,7 @@ module.exports = {
         }
 
         frame.response = {
-            pages: [mapPage(models, frame)]
+            pages: [mapper.mapPage(models, frame)]
         };
 
         debug(frame.response);
