+++ core/server/api/v2/utils/serializers/output/posts.js
@@ -1,22 +1,5 @@
 const debug = require('ghost-ignition').debug('api:v2:utils:serializers:output:posts');
-const url = require('./utils/url');
-const date = require('./utils/date');
-const utils = require('../../');
-
-const mapPost = (model, frame) => {
-    const jsonModel = model.toJSON(frame.options);
-    url.forPost(model.id, jsonModel, frame.options);
-
-    if (utils.isContentAPI(frame)) {
-        ['created_at', 'updated_at', 'published_at'].forEach((field) => {
-            if (jsonModel[field]) {
-                jsonModel[field] = date.format(jsonModel[field]);
-            }
-        });
-    }
-
-    return jsonModel;
-};
+const mapper = require('./utils/mapper');
 
 module.exports = {
     all(models, apiConfig, frame) {
@@ -29,7 +12,7 @@ module.exports = {
 
         if (models.meta) {
             frame.response = {
-                posts: models.data.map(model => mapPost(model, frame)),
+                posts: models.data.map(model => mapper.mapPost(model, frame)),
                 meta: models.meta
             };
 
@@ -38,7 +21,7 @@ module.exports = {
         }
 
         frame.response = {
-            posts: [mapPost(models, frame)]
+            posts: [mapper.mapPost(models, frame)]
         };
 
         debug(frame.response);
