+++ core/server/api/v2/utils/serializers/output/tags.js
@@ -1,22 +1,5 @@
 const debug = require('ghost-ignition').debug('api:v2:utils:serializers:output:tags');
-const url = require('./utils/url');
-const date = require('./utils/date');
-const utils = require('../../');
-
-const mapTag = (model, frame) => {
-    const jsonModel = model.toJSON(frame.options);
-    url.forTag(model.id, jsonModel, frame.options);
-
-    if (utils.isContentAPI(frame)) {
-        ['created_at', 'updated_at'].forEach((field) => {
-            if (jsonModel[field]) {
-                jsonModel[field] = date.format(jsonModel[field]);
-            }
-        });
-    }
-
-    return jsonModel;
-};
+const mapper = require('./utils/mapper');
 
 module.exports = {
     all(models, apiConfig, frame) {
@@ -28,7 +11,7 @@ module.exports = {
 
         if (models.meta) {
             frame.response = {
-                tags: models.data.map(model => mapTag(model, frame)),
+                tags: models.data.map(model => mapper.mapTag(model, frame)),
                 meta: models.meta
             };
 
@@ -36,7 +19,7 @@ module.exports = {
         }
 
         frame.response = {
-            tags: [mapTag(models, frame)]
+            tags: [mapper.mapTag(models, frame)]
         };
 
         debug(frame.response);
